transcription-script/requirements.txt
google-genai
python-dotenv
firebase-admin